+++ locators/main_page_locators.py
@@ -2,6 +2,13 @@ from selenium.webdriver.common.by import By
 
 class MainPageLocators:
 
+    cookies_accept = (By.XPATH, './/button[text()="да все привыкли"]') # Кнопка принятие куки
+    order_status_button = (By.XPATH, './/button[text()="Статус заказа"]') # Кнопка узнать статус
+    order_button = (By.XPATH, './/button[text()="Заказать"]') #Кнопка "заказать"
+    logo_yandex = (By.XPATH, './/img[@alt="Yandex"]') # Логотип яндекса
+    logo_scooter = (By.XPATH, './/img[@alt="Scooter"]') # Логотип скутера
+    yandex_search = (By.XPATH, './/iframe[@aria-label="Поиск Яндекса"]') # строка поиска в Яндексе
+    logo_dzen = (By.XPATH, "//a[@class='desktop-base-header__logoLink-aE']")
     important_questions_header = (By.XPATH, './/div[text()]="Вопросы о важном"')  # Заголовок Вопросы о важном
     important_questions = (By.XPATH, './/*[contains(@id, "accordion__heading-")]')  # Важный вопрос
     important_answers = (By.XPATH, './/*[contains(@id, "accordion__panel-")]/p')  # Важный ответ
